am62x-var-som-symphony: Fix incorrect OV5640 XTAL fixed-clock frequency
The fixed-clock node representing the external 24 MHz XTAL input to the
OV5640 camera module was incorrectly defined as 12 MHz. This frequency
information is utilized by the camera driver in order to properly set
clocks and dividers in the module. When set at 12 MHz, there is a
noticable degradation in video frame quality.

To fix this, simply correctly define the frequency at 24 MHz.

Signed-off-by: Ken Sloat <ken.s@variscite.com>